09-Multi-Agent-Capstone-Project/capstone-project/requirements.txt
openai==0.28
requests
python-dotenv
langchain==0.0.230
langchain-community
langchain-text-splitters
pymupdf
langchain-openai
langchain_huggingface
Pillow
sentence-transformers
faiss-cpu
langchain-huggingface